+++ src/main/java/com/codeborne/selenide/SelenideElement.java
@@ -1,6 +1,7 @@
 package com.codeborne.selenide;
 
 import com.codeborne.selenide.files.FileFilter;
+import com.codeborne.selenide.impl.WebElementSource;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import org.openqa.selenium.By;
 import org.openqa.selenium.NoSuchElementException;
@@ -680,6 +681,7 @@ public interface SelenideElement extends WebElement, WrapsDriver, WrapsElement, 
    * @see <a href="https://github.com/selenide/selenide/wiki/lazy-loading">Lazy loading</a>
    */
   @CheckReturnValue
+  @Nonnull
   SelenideElement find(By selector);
 
   /**
@@ -1203,7 +1205,9 @@ public interface SelenideElement extends WebElement, WrapsDriver, WrapsElement, 
   SelenideElement dragAndDropTo(String targetCssSelector, DragAndDropOptions options);
 
   /**
-   * Execute custom implemented command
+   * Execute custom implemented command (this command will not receive
+   * any arguments through {@link Command#execute(SelenideElement, WebElementSource, Object[])}
+   * when executed).
    *
    * @param command custom command
    * @return whatever the command returns (incl. null)
@@ -1211,6 +1215,20 @@ public interface SelenideElement extends WebElement, WrapsDriver, WrapsElement, 
    * @see com.codeborne.selenide.Command
    */
   <ReturnType> ReturnType execute(Command<ReturnType> command);
+
+  /**
+   * Execute custom implemented command with given timeout (this command will not receive
+   * any arguments through {@link Command#execute(SelenideElement, WebElementSource, Object[])}
+   * when executed).
+   *
+   * @param command custom command
+   * @param timeout given timeout
+   * @return whatever the command returns (incl. null)
+   * @see com.codeborne.selenide.commands.Execute
+   * @see com.codeborne.selenide.Command
+   * @since 5.24.0
+   */
+  <ReturnType> ReturnType execute(Command<ReturnType> command, Duration timeout);
 
   /**
    * Check if image is properly loaded.