+++ src/main/java/com/codeborne/selenide/conditions/And.java
@@ -9,31 +9,43 @@ import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.List;
 
-import static com.codeborne.selenide.conditions.ConditionHelpers.negateMissingElementTolerance;
 import static java.util.stream.Collectors.joining;
 
 @ParametersAreNonnullByDefault
 public class And extends Condition {
 
-  private final List<Condition> conditions;
+  private final List<? extends Condition> conditions;
   private Condition lastFailedCondition;
 
-  public And(String name, List<Condition> conditions) {
-    super(name);
+  /**
+   * Ctor.
+   *
+   * @param name       condition name
+   * @param conditions conditions list
+   * @throws IllegalArgumentException if {@code conditions} is empty
+   */
+  public And(String name, List<? extends Condition> conditions) {
+    super(name, checkedConditionsListCtorArg(conditions).stream().allMatch(Condition::missingElementSatisfiesCondition));
     this.conditions = conditions;
+  }
+
+  private static List<? extends Condition> checkedConditionsListCtorArg(List<? extends Condition> conditions) {
+    if (conditions.isEmpty()) {
+      throw new IllegalArgumentException("conditions list is empty");
+    }
+    return conditions;
   }
 
   @Nonnull
   @Override
   public Condition negate() {
-    return new Not(this, negateMissingElementTolerance(conditions));
+    return new Not(this, conditions.stream().map(Condition::negate).allMatch(Condition::missingElementSatisfiesCondition));
   }
 
   @CheckReturnValue
   @Override
   public boolean apply(Driver driver, WebElement element) {
     lastFailedCondition = null;
-
     for (Condition c : conditions) {
       if (!c.apply(driver, element)) {
         lastFailedCondition = c;
@@ -53,7 +65,6 @@ public class And extends Condition {
   @CheckReturnValue
   @Override
   public String toString() {
-    String conditionsToString = conditions.stream().map(Condition::toString).collect(joining(" and "));
-    return String.format("%s: %s", getName(), conditionsToString);
+    return getName() + ": " + conditions.stream().map(Condition::toString).collect(joining(" and "));
   }
 }