+++ statics/src/test/java/integration/FileDownloadViaProxyTest.java
@@ -9,6 +9,7 @@ import org.openqa.selenium.By;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.nio.file.Files;
 
 import static com.codeborne.selenide.Configuration.timeout;
 import static com.codeborne.selenide.DownloadOptions.using;
@@ -75,6 +76,16 @@ final class FileDownloadViaProxyTest extends IntegrationTest {
       .isEqualTo("Превед медвед!");
     assertThat(downloadedFile.getAbsolutePath())
       .startsWith(folder.getAbsolutePath());
+  }
+
+  @Test
+  void downloadsFileWithNorwayCharactersInName() throws IOException {
+    File downloadedFile = $(byText("Download file with \"ø\" in name")).download(withExtension("txt"));
+
+    assertThat(downloadedFile.getName())
+      .isEqualTo("ø-report.txt");
+    assertThat(readFileToString(downloadedFile, "UTF-8"))
+      .isEqualTo("Hello, Nørway!\n");
   }
 
   @Test
@@ -160,6 +171,22 @@ final class FileDownloadViaProxyTest extends IntegrationTest {
     File downloadedFile = $(byText("Download a PDF")).download(timeout, withExtension("pdf"));
 
     assertThat(downloadedFile.getName()).isEqualTo("minimal.pdf");
+  }
+
+  @Test
+  void downloadsPotentiallyHarmfulWindowsFiles() throws IOException {
+    File downloadedFile = $(byText("Download EXE file")).download(withExtension("exe"));
+
+    assertThat(downloadedFile.getName()).isEqualTo("tiny.exe");
+    assertThat(Files.size(downloadedFile.toPath())).isEqualTo(43);
+  }
+
+  @Test
+  void downloadsPotentiallyHarmfulMacFiles() throws IOException {
+    File downloadedFile = $(byText("Download DMG file")).download(withExtension("dmg"));
+
+    assertThat(downloadedFile.getName()).isEqualTo("tiny.dmg");
+    assertThat(Files.size(downloadedFile.toPath())).isEqualTo(43);
   }
 
   @Test