+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 5.24.2 (released 02.09.2021)
+* #1551 Fix NoClassDefFoundException for StringUtils
+
+## 5.24.1 (released 31.08.2021)
+* upgrade to WebDriverManager to 5.0.1  --  thanks to Anil Kumar Reddy Gaddam for PR #1547
+
+## 5.24.0 (released 29.08.2021)
+* #1525 add method `$.execute(Command, Duration)` for running custom commands with a custom timeout  --  thanks to Evgenii Plugatar for PR #1531
+* #1532 fix searching shadow roots inside of a web element  --  see PR #1536
+* #1527 `$.execute(Command)` and `$.execute(Command, Duration)` methods no longer pass arguments to custom command  --  thanks to Evgenii Plugatar for PR #1535
+* #1467 Avoid spam in logs when webdriver is already closed  --  see PR #1540
+* #1534 `Or` and `And` conditions work correctly with non-existent element  --  thanks to Evgenii Plugatar for PR #1539
+* `Or` and `And` conditions support PECS principle in ctor, no longer allow empty list in ctor  --  thanks to Evgenii Plugatar for PR #1542
+* #1541 removed deprecated `Condition` method `applyNull()` and renamed `CollectionCondition` `applyNull()` method  --  thanks to Evgenii Plugatar for PR #1544
+
+## 5.23.3 (released 19.08.2021)
+* #1528 fix "exe" or "dmg" file download in Chrome  -  see PR #1529
+
 ## 5.23.2 (released 03.08.2021)
 * #1508 add check `webdriver().shouldHave(numberOfWindows(N))` --  thanks to Oleg Berezhnoy for PR #1511